split up e2e tests file
